Move indoor tests to base class testing
